requirements.txt
streamlit
google-generativeai
python-Levenshtein
googlesearch-python